fix: remove internal types on publish
