+++ package.json
@@ -10,7 +10,7 @@
   },
   "scripts": {
     "build": "tsc",
-    "prepublishOnly": "tsc && rm out/tsconfig.tsbuildinfo",
+    "prepublishOnly": "tsc && rm out/utils/*.d.ts && rm out/tsconfig.tsbuildinfo",
     "test": "jest --verbose",
     "test-verbose": "tsc --sourceMap && jest --verbose --coverage --globals \"{\\\"coverage\\\":true}\"",
     "codecov": "codecov"
